requirements.txt
asgiref==3.7.2
Django==5.0.1
django-rest-authtoken==2.1.4
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
timedelta==2020.12.3
typing_extensions==4.9.0
